Changed the addr type to int to allow a graceful failure when inet_addr returns -1
git-svn-id: http://svn.research.cc.gatech.edu/kaos/atl/trunk@8762 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ atom.c
@@ -333,7 +333,7 @@ fill_hostaddr(void *addr, char *hostname)
     
     host_addr = gethostbyname(hostname);
     if (host_addr == NULL) {
-	unsigned long addr = inet_addr(hostname);
+	int addr = inet_addr(hostname);
 	if (addr == -1) {
 	    /* 
 	     *  not translatable as a hostname or
